requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
psycopg2==2.9.9
sqlparse==0.5.0
whitenoise==6.7.0

aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
async==0.6.2
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
billiard==4.2.0
bleach==6.1.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
decouple==0.0.7
dj-database-url==2.2.0
dj-rest-auth==6.0.0
django-cachalot==2.6.2
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-filter==24.2
django-phonenumber-field==7.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
drf-yasg==1.21.7
drf_api_logger==1.1.15
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
ipython==7.31.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.7
magic-filter==1.0.12
matplotlib-inline==0.1.7
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.22
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.17.2
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.6
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.4
six==1.16.0
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
